VectorAutoregressions.jl

VectorAutoregressions.jl is a package for estimating Vector Autoregressions (VARs) using Julia. The initial goal is to provide functionality comparable to the vars package in R.

Installation

At the Julia REPL:

    (@v1.x) pkg> add https://github.com/stephenbnicar/VectorAutoregressions.jl.git

Estimating a VAR

StatsBase.aic — Method

aic(v::VarEstimate)

Return Akaike's Information Criterion for VAR model v.

StatsBase.coef — Method

coef(v::VarEstimate)

Return the matrix of coefficients for VAR model v.

StatsBase.loglikelihood — Method

loglikelihood(v::VarEstimate)

Return the log-likelihood for VAR model v.

StatsBase.residuals — Method

residuals(v::VarEstimate)

Return the matrix of residuals for VAR model v.

StatsBase.stderror — Method

stderror(v::VarEstimate)

Return the standard errors for the coefficients of VAR model v.

VectorAutoregressions.VAR — Method

VAR(data, lags, constant::Bool = true, trend::Bool = false) -> VarEstimate

Estimate an unrestricted vector autoregression (VAR) using OLS.

Arguments:

data : DataFrame or TimeArray of observations on endogenous variables

lags : the number of lags

constant : Bool to indicate inclusion of intercept term

trend : Bool to indicate inclusion of a linear trend

VectorAutoregressions.hqc — Method

hqc(v::VarEstimate)

Return the Hannan-Quinn Criterion for VAR model v.

VectorAutoregressions.lagselect — Method

lagselect(data, lags, constant::Bool = true,
    trend::Bool = false) -> LagSelectionCriteria

Calculate AIC, SIC, and HQC lag selection criteria for an unrestricted VAR.

Arguments:

data : DataFrame or TimeArray of observations on endogenous variables

lags : maximum number of lags

constant : boolean to indicate inclusion of intercept term (default is true)

trend : boolean to indicate inclusion of a linear trend

VectorAutoregressions.sic — Method

sic(v::VarEstimate)

Return the Schwarz (Bayesian) Information Criterion for VAR model v.

VectorAutoregressions.LagSelectionCriteria — Type

LagSelectionCriteria

Fields:

maxlag : Int

table : DataFrame

selction : Dict

VectorAutoregressions.companion — Method

companion(A, p)

Construct the companion matrix for VAR coefficient matrix A, where A has dimensions (Kp × K).