VectorAutoregressions.jl

VectorAutoregressions.jl is a package for estimating Vector Autoregressions (VARs) using Julia. The initial goal is to provide functionality comparable to the vars package in R.

Estimating a VAR

VectorAutoregressions.VAR — Function

VAR(data, lags; constant = true, trend = false) -> VarEstimate

Estimate an unrestricted vector autoregression (VAR) using OLS.

Arguments

data : DataFrame or TimeArray of observations on endogenous variables

lags::Int : the number of lags

constant::Bool = true : include an intercept term

trend::Bool = false : include a linear trend

VectorAutoregressions.VarEstimate — Type

VarEstimate

Fields

Y::Union{DataFrame,TimeArray}

X::Union{DataFrame,TimeArray,Nothing}

ynames::Array{String}

xnames::Array{String}

lags::Int

constant::Bool

trend::Bool

obs::Int

Z::Matrix

B::Matrix

seB::Matrix

U::Matrix

ΣU::Matrix

Yhat::Matrix

Diagnostics

VectorAutoregressions.lagselect — Function

lagselect(data, maxlag; constant = true, trend = false) -> VarLagSelection

Calculate AIC, SIC, and HQC lag selection criteria for an unrestricted VAR.

Arguments

data : DataFrame or TimeArray of observations on endogenous variables

maxlag::Int : the maximum number of lags to estimate

constant::Bool = true : include an intercept term

trend::Bool = false : include a linear trend

VectorAutoregressions.VarLagSelection — Type

VarLagSelection

Fields

maxlag::Int

table::DataFrame

selection::Dict

VectorAutoregressions.checkstable — Function

checkstable(v::VarEstimate) -> VarStabilityCheck

Check the stability of the estimated VAR model v.

VectorAutoregressions.VarStabilityCheck — Type

VarStabilityCheck

Fields

isstable::Bool

eigenvals::Array{Number}

eigenmod::Array{Float64}

VectorAutoregressions.portmanteau_test — Function

portmanteau_test(v::VarEstimate, h::Int)

Conduct a multivariate portmanteau test for residual autocorrelation up to lag h for VAR model v. Implements the adjusted test described on p.171 of Lutkepohl (2006).

StatsBase.loglikelihood — Function

loglikelihood(v::VarEstimate)

Return the log-likelihood for VAR model v.

StatsBase.aic — Function

aic(v::VarEstimate)

Return Akaike's Information Criterion for VAR model v.

VectorAutoregressions.sic — Function

sic(v::VarEstimate)

Return the Schwarz (Bayesian) Information Criterion for VAR model v.

VectorAutoregressions.hqc — Function

hqc(v::VarEstimate)

Return the Hannan-Quinn Criterion for VAR model v.

Other

StatsBase.coef — Function

coef(v::VarEstimate)

Return the matrix of coefficients for VAR model v.

StatsBase.stderror — Function

stderror(v::VarEstimate)

Return the standard errors for the coefficients of VAR model v.

StatsBase.residuals — Function

residuals(v::VarEstimate)

Return the matrix of residuals for VAR model v.

StatsBase.fitted — Function

fitted(v::VarEstimate)

Return the fitted values for VAR model v.